Created the VC++ project. Rewrote the scene framework; the process is now much smoother.
scene_list.h holds a list of the scenes currently usable; both SceneManager and Scene depend on it.

Scene is the abstract base class for all of the scenes in the program. The Scene class allows for polymorphism in SceneManager. It has the necessities built in so the user can copy and change whatever they want. It has the pointer to the screen and the index of the next scene as private static members.

SceneManager is no longer spread across two classes, instead existing as a single discreet unit; this of course, allows the user to actually see it's internal workings. It no longer has it's own "next scene" member, rather depending on the Scene's static member.

I've created "scene generator.lua", a lua script designed to output valid, compilable scenes using only a few parameters. It's really time consuming writing these from scratch, and the basic outline is always the same, so this was a natural solution. I've included "SceneExample" as an example output, and threaded it through the framework.

I've included "dir-build.bat" as a solution to quickly copy the files into a usable directory. Right now it only copies the framework and the scene generator. When the Codebase is finished, it'll copy the framework, all of the header files and the .lib binary into the usable directory tree, which I can just copy to the C: drive.

I've added some text files for simple documentation, but I've also added Layout.docx to decide on the layout of the Codebase in general. The file "scene framework layout.class.violet" is opened using "com.horstmann.violet-0.21.1.jar", which can be found on the net.