Created BBox
Are you ready for a comedy of errors?

I keep commiting new modules and forgetting to add the comment headers. In
BBox, I did it again, so I added the headers and amended the commit.
But... I forgot to add the files to the staging area, so it was an empty
commit.

Funny huh?