Created Animator, sorted out class list
Animator is a number crunching module, which means there is no actual
graphics or rendering code present. It's only function is to signal what
frame of a sprite sheet to draw at a specidfied time. It is independant of
SDL.

A rough class list is in TODO.txt; when I've finished what is on this
list, hopefully it can be considered "ready"