Created project "Hello World" to demonstrate how to use the codebase.
There are two child classes of scene, MainMenu and HelloWorld. The game
starts by loading "config.cfg", and passing the ConfigUtility to the newly
created scenes. MainMenu is created first, and shows a pair of buttons
which, when clicked, either switch to the HelloWorld scene, or exit the
game. The second loads a player and flower actor. the flower is simply
animated, while the player can move around the screen.

Right now, there are only rough menu buttons and Actors. These can be expanded and modified as needed.